TransactionModule_2

+++ Transactions/Transactions.Application.Contract/Transaction/Command/ITransactionCommands.cs
@@ -6,7 +6,7 @@ namespace Transactions.Application.Contract.Transaction.Command
     public interface ITransactionCommands
     {
         Task<OperationResult> CreateAsync(CreateTransacionCommandModel commnad);
-        Task<OperationResult> Payment(TransactionStatus status,int id,string refid); 
+        Task<OperationResult> Payment(TransactionStatus status,long id,string refid); 
     }
 
 }
